+++ packages/box/src/component.tsx
@@ -1,7 +1,7 @@
 import { classNames } from '@chbphone55/classnames';
 import { box as ccBox } from '@warp-ds/css/component-classes';
 import React from 'react';
-import { BoxProps } from './props';
+import { BoxProps } from './props.js';
 
 export function Box(props: BoxProps) {
   const {