+++ packages/pill/src/component.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { classNames } from '@chbphone55/classnames';
-import { PillProps } from '.';
+import { PillProps } from './props.js';
 import { pill as ccPill } from '@warp-ds/css/component-classes';
 import { i18n } from '@lingui/core';
 import { messages as nbMessages} from './locales/nb/messages.mjs';
 import { messages as enMessages} from './locales/en/messages.mjs';
 import { messages as fiMessages} from './locales/fi/messages.mjs';
-import { activateI18n } from '../../i18n';
+import { activateI18n } from '../../i18n.js';
 
 export function Pill(props: PillProps) {
   activateI18n(enMessages, nbMessages, fiMessages);