+++ packages/slider/src/index.ts
@@ -1,2 +1,2 @@
-export { Slider } from './component';
-export type { SliderProps } from './props';
+export { Slider } from './component.js';
+export type { SliderProps } from './props.js';