+++ packages/tabs/src/component-tabs.tsx
@@ -7,8 +7,8 @@ import React, {
 } from 'react';
 import { classNames } from '@chbphone55/classnames';
 import { gridLayout, tabs as ccTabs } from '@warp-ds/css/component-classes';
-import { debounce } from './utils';
-import type { TabsProps } from './props';
+import { debounce } from './utils.js';
+import type { TabsProps } from './props.js';
 
 const setup = (
   { className, children, onClick, active, ...rest }: any,